robot_workcell_agent/requirements.txt
# Core Agent Framework
pydantic-ai[anthropic,openai]>=0.0.30
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Skill System
pyyaml>=6.0

# LLM Providers
anthropic>=0.39.0
openai>=1.0.0
httpx>=0.27.0

# CLI & UI
rich>=13.0.0
streamlit>=1.30.0

# Scientific / Optimization
numpy>=1.21.0
pyswarms>=1.3.0

# Genesis Physics Engine (install separately: pip install genesis-world)
genesis-world>=0.1.0

# Environment variables
python-dotenv>=1.0.0
